Move GetConsolidateAuthenticationPolicy under 'version' folder to separate logic between API versions
